method findAllByVehicleId to JourneyDaoImpl;
adjusted getTotalKilometrage method in VehicleDaoImpl;
adjusted failing test;

+++ fleet-management-persistence/src/main/java/cz/fi/muni/pa165/dao/interfaces/JourneyDao.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public interface JourneyDao extends Dao<Journey, Long> {
 
-    List<Journey> findAllByVehicleId(double vehicleId);
+    List<Journey> findAllByVehicleId(long vehicleId);
     
    /**
     * find all journeys for given employee
